Download and format data for RSV dynamic transmission model

Download and format data for RSV dynamic transmission model

Usage

get_data(state_or_county, state_abbr, county_name)

Arguments

state_or_county

describe if you are modeling for a "state" or a "county"

state_abbr

The two letter state abbreviation (you need this even if you are modeling a county)

county_name

The name of the county (if modeling a state list NULL here)

Value

a list of values

Examples

data_california = get_data(state_or_county="state",state_abbr="CA",county_name=NULL)
#> Using the Vintage 2022 Population Estimates
#> Warning: For post-2020 Census estimates, `get_estimates()` now uses the `vintage` argument to specify the PEP vintage, and the `year` argument to isolate a year within that vintage.
#> ! This may be a breaking change in your code
#> ! Omitting `vintage` may lead to incorrect or unexpected results.
#> Using the Vintage 2022 Population Estimates
data_kingcounty = get_data(state_or_county="county",state_abbr="WA",county_name="King")
#> Using the Vintage 2022 Population Estimates
#> Warning: For post-2020 Census estimates, `get_estimates()` now uses the `vintage` argument to specify the PEP vintage, and the `year` argument to isolate a year within that vintage.
#> ! This may be a breaking change in your code
#> ! Omitting `vintage` may lead to incorrect or unexpected results.
#> Using the Vintage 2022 Population Estimates